[Refactor] move utils
